% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.werm_fit}
\alias{summary.werm_fit}
\title{Summarise a fitted WERM model}
\usage{
\method{summary}{werm_fit}(object, ...)
}
\arguments{
\item{object}{A \code{werm_fit} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{object}. Called for its printed side-effect.
}
\description{
Summarise a fitted WERM model
}
