% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{weightederm_configure_python}
\alias{weightederm_configure_python}
\title{Configure which Python environment weightederm uses}
\usage{
weightederm_configure_python(python, required = TRUE)
}
\arguments{
\item{python}{Path to the Python binary or virtual-environment directory.}

\item{required}{Passed to \code{\link[reticulate:use_python]{reticulate::use_python()}}.}
}
\description{
A thin wrapper around \code{\link[reticulate:use_python]{reticulate::use_python()}} that must be called
\strong{before} any \verb{werm_*} function if the default Python does not have
\code{weightederm} installed.
}
\examples{
\dontrun{
weightederm_configure_python("~/.venv/bin/python")
}

}
